Change protocol-host package to match its name
